PR Best Practices Update
The section on PR review hasn't been updated in 6 years and it's something we do every day! This is just as starting point that includes things from engineers that have come up over the last few months. Happy to discuss all of this here and make further updates. In addition to a general rewording of this I've added:
- Recommendation to create CODEOWNERS files to make it easier to request review from entire teams (and make it the default action)
- Added policy for re-requesting review so that PRs don't get lost after an initial review
- Policies around PR complexity (should have a substantive review, but also can request a smaller PR)