TAD Timeline update
After many conversations with engineers about length of the TAD drafting process I'd like to float a idealized version of the process here. This isn't intended to be prescriptive but to provide guidence on how someone _could_ move through the drafting process. Any and all feedback on this is more than welcome.

I've also made another change and shifted two situations from MUST to SHOULD for when a TAD is needed. These are the clauses for when work will be completed by multiple engineers or over mulitple sprints. While sometimes useful we are not regularly following this practice and it seems to be an uneccsary requirement.