Onboarding Re-Write
Our onboarding documentation was out of date and focused largely on adding people to accounts. This is misleading and neglects to mention the main goal of our _actual_ onboarding process: To welcome people to the team!

This rewrite attempts to center that more and update the account list to better reflect what is used and important in NYPL Digital (e.g. Less Jenkins, More GitHub)